Fix FoodItem entity and repository

+++ OrderApp/src/main/java/com/criskong/OrderApp/Entity/FoodItem.java
@@ -21,7 +21,7 @@ public class FoodItem {
     //Genera automaticamente l'id del prodotto
     @Id
     @GeneratedValue(strategy=GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String name;
     private double price;
 
@@ -30,16 +30,16 @@ public class FoodItem {
 
     }
     //costruttore default
-    public FoodItem(int id, String name, double price){
+    public FoodItem(Long id, String name, double price){
             this.id = id;
             this.name = name;
             this.price = price;
     }
     
-    public int getId(){
+    public Long getId(){
         return id;
     }
-    public void setId(int id){
+    public void setId(Long id){
         this.id = id;
     }
